[Refact] 움직임 정측을 서로 다른 파일로 관리
